+++ src/main/kotlin/it/czerwinski/intellij/wavefront/lang/MtlSyntaxHighlighter.kt
@@ -45,6 +45,8 @@ class MtlSyntaxHighlighter : SyntaxHighlighterBase() {
             MtlTypes.SPECULAR_EXPONENT_KEYWORD,
             MtlTypes.SHARPNESS_KEYWORD,
             MtlTypes.OPTICAL_DENSITY_KEYWORD,
+            MtlTypes.ROUGHNESS_KEYWORD,
+            MtlTypes.METALNESS_KEYWORD,
             MtlTypes.AMBIENT_COLOR_MAP_KEYWORD,
             MtlTypes.DIFFUSE_COLOR_MAP_KEYWORD,
             MtlTypes.SPECULAR_COLOR_MAP_KEYWORD,
@@ -52,6 +54,9 @@ class MtlSyntaxHighlighter : SyntaxHighlighterBase() {
             MtlTypes.DISSOLVE_MAP_KEYWORD,
             MtlTypes.DISPLACEMENT_MAP_KEYWORD,
             MtlTypes.STENCIL_DECAL_MAP_KEYWORD,
+            MtlTypes.ROUGHNESS_MAP_KEYWORD,
+            MtlTypes.METALNESS_MAP_KEYWORD,
+            MtlTypes.NORMAL_MAP_KEYWORD,
             MtlTypes.BUMP_MAP_KEYWORD,
             MtlTypes.REFLECTION_MAP_KEYWORD ->
                 arrayOf(ATTR_KEYWORD)